TLS User Sessions Demo
This project is a proof-of-concept that demonstrates a method of
handling a user session once a user is authenticated using TLS mutual
authentication with x509 certificates; while maintaining a user friendly
experience on the client.  This is an alternative to traditonal session
management built on top of the HTTP that typically involve session
tokens inserted into headers or provided via cookies.